// lib/supabase/database.types.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "14.1"
  }
  public: {
    Tables: {
// ...
      rewards: {
        Row: {
          created_at: string | null
          description: string | null
          id: string
          is_active: boolean | null
          is_coming_soon: boolean | null
          name: string
          points_required: number
          reward_type: string
          stock_quantity: number | null
          updated_at: string | null
          value_amount: number
        }
        Insert: {
          created_at?: string | null
          description?: string | null
          id?: string
          is_active?: boolean | null
          is_coming_soon?: boolean | null
          name: string
          points_required: number
          reward_type: string
          stock_quantity?: number | null
          updated_at?: string | null
          value_amount: number
        }
        Update: {
          created_at?: string | null
          description?: string | null
          id?: string
          is_active?: boolean | null
          is_coming_soon?: boolean | null
          name?: string
          points_required?: number
          reward_type?: string
          stock_quantity?: number | null
          updated_at?: string | null
          value_amount?: number
        }
        Relationships: []
      }
// ...
    }
// ...
  }
}